+++ dataactvalidator/config/detachedAwardFields.csv
@@ -34,7 +34,7 @@ periodofperformancecurrentenddate,period_of_performance_curr,FALSE,str,8,"D3, D4
 periodofperformancestartdate,period_of_performance_star,FALSE,str,8,"D3, D18",
 primaryplaceofperformancecode,place_of_performance_code,TRUE,str,7,"D1, D14, D15",
 primaryplaceofperformancecongressionaldistrict,place_of_performance_congr,FALSE,str,2,,TRUE
-primaryplaceofperformancecountrycode,place_of_perform_country_c,TRUE,str,3,D1,
+primaryplaceofperformancecountrycode,place_of_perform_country_c,FALSE,str,3,D1,
 primaryplaceofperformanceforeignlocationdescription,place_of_performance_forei,FALSE,str,100,D17,
 primaryplaceofperformancezip+4,place_of_performance_zip4a,FALSE,str,10,"D2, D16",
 recordtype,record_type,TRUE,int,1,D1,
