+++ dataactvalidator/config/sqlrules/fabs24_detached_award_financial_assistance_2.sql
@@ -1,0 +1,9 @@
+-- PrimaryPlaceOfPerformanceCountryCode must be blank for record type 3.
+SELECT
+    dafa.row_number,
+    dafa.record_type,
+    dafa.place_of_perform_country_c
+FROM detached_award_financial_assistance AS dafa
+WHERE submission_id={0}
+    AND record_type = 3
+    AND COALESCE(place_of_perform_country_c, '') <> ''
